Adds ViewModel for window
